Rename functions file
