+++ packages/cli/lib/getFunctionArrays.js
